Clean up bytecode/lib.rs
